feat!: Anisotropic mode for default building
When asked for a mesh, the anistropic mode is now available. It works
by, firstly, building an isotropic mesh, then deforming last dimension
(y in 2D, z in 3D). Roughly speaking, the components of each point is
replaced by its exponential (that is z_0->exp(z_0)). Finally, the mesh
is rescaled to ensure that each dimension has the same size